vm_tools: Use HiDPI for X forwarding by default.
This change makes sommerlier detect the default scale factor
and use it as reference scale when X forwarding is enabled.

The user can still adjust the scale using the --scale flag. For
example, --scale=0.5 reduces the density to half of the default
scale factor.

BUG=none
TEST=atom, android studio, etc. use HiDPI without --scale=2

Change-Id: Ifab07651508eee63034d5861a87ce1bd4332520e
Reviewed-on: https://chromium-review.googlesource.com/1001708
Commit-Ready: David Reveman <reveman@chromium.org>
Tested-by: David Reveman <reveman@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>