vm_tools: sommelier: handle window gravity correctly
First, we remove any weird gravity when we start managing a window
as that seems to be a common practice in other WMs and allows the
application to detect that gravity changed.

This also makes the resize code more robust against problems with
window gravity. Basically, we don't assume that it's northwest as
a misbehaving client can always try to change this back to
something else.

BUG=chromium:843570
TEST=resize of libreoffice windows work correctly

Change-Id: Ie13b30c1e785151d32cd2d66ec11e1954b46c1fa
Reviewed-on: https://chromium-review.googlesource.com/1061945
Commit-Ready: David Reveman <reveman@chromium.org>
Tested-by: David Reveman <reveman@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>