vm_tools: sommelier: Pass xauth flag to Xwayland
This is to enable apps like synaptic to work properly where they elevate
their privileges to root and then need to connect to the X server. We
setup the .Xauthority file in the service file for sommelier-x. Then we
pass the path to the .Xauthority file to Xwayland.

This also adds a xasprintf method which removes the snprintf usage and
cleans up some parts of the code.

BUG=chromium:876846
TEST=Verified synaptic-pkexec works with both X servers (0 and 1)

Change-Id: I538b6d5a72c7c268ffc3229f9903adaa7620a6ed
Reviewed-on: https://chromium-review.googlesource.com/1279105
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>