vm_tools: sommelier: Set LIBGL_DRIVERS_PATH for Xwayland.
Xwayland is typically built from the chroot and uses lddtree to find the
appropriate GL libraries.  Mesa does an explicit dlopen of the DRI
drivers and needs LIBGL_DRIVERS_PATH set to find the libraries linked
against the libc that Xwayland was built with.

BUG=chromium:837073
TEST=xterm, glxgears

Change-Id: I10c1ec40cba7429db5b400d370ae371297c0190a
Reviewed-on: https://chromium-review.googlesource.com/1403896
Commit-Ready: David Riley <davidriley@chromium.org>
Tested-by: David Riley <davidriley@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>