vm_tools: sommelier: Fix wayland data transfers
Wayland data transfers using wl_data_offer::receive and
wl_data_source::send only ever worked by chance. The transfer code would
delete the read event source when it ran out of data with the intention
that the writer would finish writing and then clean up, but the writer
could not be enabled for events at that time, so it would never wake up
and the transfer would appear to fail.

The specific application response to this varied, which made uncovering
this issue more difficult. Some versions of GTK3 close their end of the
pipe before sommelier does, this would cause sommelier to get a
WL_EVENT_ERROR on the write source, prompting it to end the transfer.
Others would ignore the uncompleted transfer, prompting bug reports of
wayland pasting not working correctly. wl-clipboard would print out all
data as it arrived which, combined with the gtk3 bug concealing this
one, appeared to be a bug in wl-clipboard.

TEST=manually tested
BUG=chromium:986835

Change-Id: I788bfc4d9b1e13221ad8b56e826e550184c3a93e
Reviewed-on: https://chromium-review.googlesource.com/1722645
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Ready: Fergus Dall <sidereal@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: David Reveman <reveman@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>