vm_tools: Fix FD double-close in sommelier
Both the wl_shm_pool and the sl_mmap claimed to own the FD associated
with a memory-mapped buffer, and both would close try to close it.
Previously this was fixed by dup()ing the FD, but that leads to a large
number of open FDs (one for every buffer) at runtime.

Instead we remove the fd handle from the sl_mmap in the case of
client-side buffers, allowing the fd to be closed when the pool is
destroyed.

Host-side buffers still have the FD handle and should continue to, since
they need to map GPU memory or other wacky things.

BUG=chromium:1101118
TEST=for i in {1..16}; do freecad & done; emacs

Change-Id: Id0c7e1991cd955d925749626fa10bedcf99fa545
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2299667
Tested-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>