vm_tools: sommelier: do not commit if there's no window change
We add a condition in sommelier that prevents a pending configuration
from being committed if there is nothing to commit.

This was causing a problem where the CPU would be thrashed because exo
continually re-configures buffers until the client maps them every
time they commit, and sommelier would commit a buffer every time it was
configured, even if it was unmapped and therefore the configuration was a
no-op.

BUG=b:181077580, b:179285243
TEST=see b/179285243 for details

Change-Id: Ibdc659e6b6008afaca6f472fe0fd6c687a43590d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2781172
Tested-by: Nic Hollingum <hollingum@google.com>
Auto-Submit: Nic Hollingum <hollingum@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: David Riley <davidriley@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>