vm_tools: sommelier: Implement wl_surface::damage_buffer.
For now this is behind a flag, --support-damage-buffer.

For simplicity it's implemented in terms of wl_surface::damage, so
the host (such as Exo) doesn't need to implement
wl_surface::damage_buffer itself. The implementation could be
streamlined if wl_surface::damage_buffer is implemented on the
host in future, however.

Tested using weston-simple-damage as a conformance test. In a
Chrome OS Linux VM, it now displays the same visual results when
sending wl_surface::damage_buffer vs sending wl_surface::damage
(toggled via --use-damage-buffer flag).

The following bugs were observed, but are not regressions as they
occur with or without --use-damage-buffer:

* Red pixels appear with the following command, indicating an error
  in applying the viewport (for any scale != 1):
    weston-simple-damage --use-viewport --scale=2
* Any of --transform={90,180,270} causes smearing. All "flipped"
  transforms, and the default transform, appear to work correctly.

BUG=chromium:1200726,b:171844178
TEST=sommelier --support-damage-buffer --scale=0.5 weston-simple-damage --scale=3 --use-damage-buffer; sommelier --support-damage-buffer weston-simple-damage --use-viewport --use-damage-buffer

Change-Id: Id2279454786c2755d6102edd766326f56f66d0ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2879800
Commit-Queue: Chloe Pelling <cpelling@google.com>
Tested-by: Chloe Pelling <cpelling@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>