vm_tools: sommelier: Make --peer-cmd-prefix default to ""
Currently when the parent sommelier runs a child sommelier process it
tries to exec ld-linux on the ELF binary directly (with the path taken
from argv[0]). Historically this worked, but it bypasses the script we
get from lddtree and can drift out of sync.

Also, when glibc was upreved recently ld-linux learned the --argv0
option which seems to now be used to point argv[0] at the lddtree
script. This means the current way is trying to run the dynamic linker
on a script, which fails. Fix this by removing the default peer
command prefix so we just exec the script directly. I'm leaving in the
command line option and env variable because it might be useful in
debugging.

BUG=chromium:1204368
TEST=gedit

Change-Id: I65c669351d0eca47fc0eae840dac7d147a66109a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2866287
Tested-by: Fergus Dall <sidereal@google.com>
Auto-Submit: Fergus Dall <sidereal@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>