vm_tools: sommelier: Clarify sl_handle_selection_fd_readable
The allocation code here is not-quite-wrong, but is very confusing and
has lead me to try and fix imagined bugs in it on two separate
occasions. Re-write it a bit and add a comment to make it clearer what
is happening and why.

Incidentally, this fixes a real bug where we re-allocate the buffer if
the amount of data read turns out to be small. This should not affect
correctness (as wl_array_add adds the newly allocated region to the
end of the *used* part of the array, not the *allocated* part, so any
newly read data is contiguous with the existing data), but is slightly
more efficient.

BUG=chromium:1205764
TEST=CQ

Change-Id: I8b4c7bc13df8825556c1d798919bdf30a41f547c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2879282
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Nic Hollingum <hollingum@google.com>