vm_tools: sommelier: Add a wrapper main
To unit test sommelier we need to build the part-under-test as a
library without a main function, and then link that into a test
binary. Right now sommelier doesn't have the internal separation that
would be required to make this meaningful, so instead just
move *everything* into the "library" including the main function.

To make this link, we can rename "main" to "real_main" and add a
trivial wrapper main over it when linking as an executable.

BUG=b:190088895
TEST=USE="asan" emerge-tatl sommelier

Change-Id: I7c214d831a8181d7aa32eb3e742a6f936123de28
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2972326
Tested-by: Fergus Dall <sidereal@google.com>
Auto-Submit: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>