vm_tools/sommelier: use page aligned sizes for virtgpu
The DRM subsystem requires all allocations and memory mappings
to be page-aligned.  The host does this for the VirtGpuChannel
already, though that information must be propagated to the common
layer.

In addition, we must specify a format even when we just want to
allocate a bunch of bytes for wayland-shm.  This is traditionally
done via DRM_FORMAT_R8 with the GBM + gralloc APIs.

BUG=b:150239451
TEST=local testing with virtgpu + virtwl channels

Change-Id: I681f30eebe8343e2c8344266cb1bb4707ca59f84
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3021938
Commit-Queue: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>