vm_tools/sommelier: allow other operations besides receive
The VirtGpuChannel operates slightly differently than the
VirtWaylandChannel.  With the VirtWaylandChannel, it is possible
to use the write(2) and read(2) system calls on pipe file
descriptors created by virtwl and the buffered data gets sent
directly to the host.

With the VirtGpuChannel, virtgpu does not create pipe file
descriptors and therefore must proxy the operation.  So for every
write end received by the host, we need to:

1) Create a read and write end in the guest
2) Send to write end to the guest client
3) Poll the read end in Sommelier

There's a similar trick used for read ends.  As such, the receive
operation becomes more generic.  It is now known as
"handle_channel_event" to better capture it's expanded scope.

BUG=b:150239451
TEST=compile and run

Change-Id: I2f843e0ae0306c65f41710ea669d5fabfbf48363
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3054320
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Gurchetan Singh <gurchetansingh@chromium.org>