vm_tools: Fix Sommelier fullscreen order-of-init.
If sl_window_update() was first called after an X11 fullscreen
request was received, the fullscreen request would be ignored.
Save the window's desired fullscreen state and set it in
sl_window_update(), similar to how maximizing already works.

This may have the side-effect of making the fullscreen state
"stickier", in case the window is kicked out of fullscreen for
some reason.

I added a few unit tests for this. It would be ideal to have
coverage of sl_window_update() too, but that would require setting
up more pieces of Sommelier's Wayland server impl in the test.
Currently many of those pieces are initialized in real_main(),
so we can't use them in the test without further refactoring.

BUG=b:202879129
TEST=see bug for repro instructions

Cq-Depend: chromium:3238469
Change-Id: I21358a34b4dc402e691aad140dfb9b242227412e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3225040
Tested-by: Chloe Pelling <cpelling@google.com>
Commit-Queue: Chloe Pelling <cpelling@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>