vm_tools/sommelier: support virtgpu dummy event
The current virtgpu implementation of poll(..) drops events
when VIRTGPU_CONTEXT_PARAM_POLL_RINGS_MASK is enabled (otherwise
it's like a normal DRM driver).

This is because paravirtualized userspaces typically receive
responses in a buffer of type BLOB_MEM_GUEST.

To be in line with other DRM drivers and avoid specialized behavior,
it is possible to define a dummy event for virtgpu.  Paravirtualized
userspace will now have to call read(..) on the DRM fd to receive the
dummy event.

This isn't relevant to Crostini/Borealis yet since it is still on
kernel 5.10 and still uses virtio-wl.

BUG=b:207409779
TEST=run some Wayland apps + kernel changes applied

Change-Id: Ibb41310907cfe890ee725facc6008e0f39fd3909
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3296610
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Commit-Queue: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Nicholas Verne <nverne@chromium.org>