vm_tools: sommelier: Adjust scale factors if needed
This patch fixes the glbench encountered in the bug below.
On request, a check will be performed to see if the desired
pixel dimensions of the window can be converted cleanly into
the logical space.

If it cannot, the surface level scale factor override will be used
to make this match.

The effect of this patch is that the surface buffers will match what
is requested, but there will be some slight scaling done on the
presented surface within the logical space.

BUG=b:217976094
TEST=Ran glbench and some games with different scaling settings in direct/non-direct mode.

Change-Id: Ib0c9c813e6b1622ecbef389e2ee336a235c0f2c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3833450
Reviewed-by: Chloe Pelling <cpelling@google.com>
Commit-Queue: Isaac Bosompem <mrisaacb@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Tested-by: Isaac Bosompem <mrisaacb@google.com>