vm_tools: sommelier: Add DRM mmap code
This patch adds support for mapping a DRM prime buffer within Sommelier.

For the XShape support to work within Sommelier, we will need to read
the image coming in from the guest. The guest can supply an image as a
SHM memory buffer or a DRM buffer. To read the DRM buffer we will need
to map it, which is what this patch enables.

BUG=b:223232234
TEST=Ensured no crashes when using both regular and shaped applications

Change-Id: Id4561db7765845e37ead6496410438b665a7afd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3880724
Reviewed-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Isaac Bosompem <mrisaacb@google.com>
Tested-by: Isaac Bosompem <mrisaacb@google.com>
Reviewed-by: Chia-I Wu <olv@google.com>