vm_tools: Set surface scale per output in sommelier
Currently the scaling on the internal monitor is always used. This causes windows on the external monitor to change in size depending on the internal monitor scaling.

Fixes scaling issues in ApeOut on the external monitor but it still seems to have scaling issues on the internal display when a monitor is connected.

Cq-Depend: chromium:3974507
BUG=b:191820689, b:208720580
TEST=Ran Borealis and Crostini with external monitor

Change-Id: I8a35d71885791dd797468c643da669e52bff258d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3960650
Tested-by: Lucy Qu <lqu@google.com>
Reviewed-by: Chloe Pelling <cpelling@google.com>
Commit-Queue: Lucy Qu <lqu@google.com>