vm_tools: Add WeakResourcePtr type to sommelier
Sommelier uses wl_resource_add_destroy_listener in its zwp_text_input,
wl_seat/wl_pointer/wl_touch support to maintain a weak pointer to
wayland objects. As the logic involved is a bit tricky, this CL moves it
into a new WeakResourcePtr class to make it easier for future clients to
implement similar behaviour and reduce code duplication.

For now we only move the text_input usage. The behaviour for the input
types is a bit more complex as we need to invoke more logic when the
observed object is destroyed. This will require adding a callback to the
WeakResourcePtr type, and will be done in a separate CL.

BUG=b:222629444
TEST=Manually test deleting surfaces associated with text_input objects
and ensure no crashes

Change-Id: Ie75505cdd65e3de5e1336eb840c103f7fa978e0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3974507
Reviewed-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Timothy Loh <timloh@chromium.org>
Reviewed-by: Lucy Qu <lqu@google.com>
Tested-by: Timothy Loh <timloh@chromium.org>