vm_tools: sommelier: wait for fences attached to dma-buf
When a fence is attached to a dma-buf, wait for it to signal before sending
dma-buf to the host. For virtgpu fences keep the existing wait or when explicit
fences are enabled proceed with sending them to the host along with dma-buf.

BUG=b:279966749
TEST=run games via virgl and PCI passthrough.

Change-Id: Ieee91d8627b70655e35f942157afc42556e40cc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4507419
Auto-Submit: Dominik Behr <dbehr@chromium.org>
Reviewed-by: David Riley <davidriley@chromium.org>
Tested-by: Ryan Neph <ryanneph@google.com>
Tested-by: Dominik Behr <dbehr@chromium.org>
Commit-Queue: Ryan Neph <ryanneph@google.com>
Reviewed-by: Ryan Neph <ryanneph@google.com>