vm_tools: sommelier: Define quirks config.
We create a protobuf definition which allows matching windows based on
conditions (for now, just their STEAM_GAME property) and enabling
specific features if those conditions match.

This CL defines and compiles the protobuf, and adds tests for the
initial window-matching functionality. Integrating this into production
Sommelier code is to follow in a later CL.

This adds a dependency on protobuf, so it's optional and disabled by
default. For now only Borealis intends to use it, so it's also not
referenced in BUILD.gn.

BUG=b:296327135
TEST=run unit tests: meson setup builddir -Dquirks=true; pushd builddir; ninja && ./sommelier_test; popd

Change-Id: Ic15808e333517fa1bc603bed0107999a104ab344
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4886082
Reviewed-by: Nic Hollingum <hollingum@google.com>
Tested-by: Chloe Pelling <cpelling@google.com>
Auto-Submit: Chloe Pelling <cpelling@google.com>
Commit-Queue: Chloe Pelling <cpelling@google.com>