vm_tools: sommelier: Add quirk to unset X11 maximize properties.
Some games behave poorly when they are in the X11 "maximized" states
(specifically if `_NET_WM_STATE_MAXIMIZED_HORZ/VERT` are set). This
adds a quirk to disable setting these properties in X11, but allows
for Exo/Sommelier to resize the window as if in a maximized state.

There are two cases that are handled:
- If the game sets the properties (tries to maximize itself), a
call to `xdg_toplevel_set_maximized()` is made but then the X11
properties are unset in the next `sl_internal_toplevel_configure`
call. Exo then sets the window state to maximized and resizes
the window.
- If exo tries to maximize the game, it maintains its own
maximized state but does not set the properties for X11. The game
window is resized as if maximized from the X11 point of view.

BUG=b:324581354
TEST=test with quirk enabled in Cookie Clicker

Change-Id: I9e9a90ea5726c2b210a6bd77837dbf2b7d3381b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5458780
Tested-by: Arjun Srinivasan <asriniva@google.com>
Reviewed-by: Chloe Pelling <cpelling@google.com>
Commit-Queue: Arjun Srinivasan <asriniva@google.com>