vm_tools: sommelier: set log-level from cmdline
New style logging uses the syntax `LOG(INFO) << "message";` to indicate
the logging verbosity. That verbosity is tested prior to visible output,
but the selection of which levels will be shown is only controllable at
build-time via '-DLOG_LEVEL=-1'.

There's no practical reason to require such. The test for visibility
_may_ get optimized out by compiler for statically-defined LOG_LEVEL, but
it is a trivial cost not worth optimizing in the general case. Other
tricks can be used to solve that if it becomes an issue.

Now it can be set with --log-level=<int>, with the default set in the
old way with -DLOG_LEVEL=<int>, or defaulting to LOG_LEVEL_INFO (0) when
not defined.

BUG=b:339048928
TEST=sommelier --log-level=-1 ...

Change-Id: I2de93e6c11ccba89022a4aa460c082644328e800
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5517861
Tested-by: Ryan Neph <ryanneph@google.com>
Reviewed-by: Chloe Pelling <cpelling@google.com>
Auto-Submit: Ryan Neph <ryanneph@google.com>
Reviewed-by: Arjun Srinivasan <asriniva@google.com>
Commit-Queue: Ryan Neph <ryanneph@google.com>