vm_tools: sommelier: Change containerized aspect ratio logic.
The current logic for containerized windows with a viewport override
(no randr emulation) attempts to set the aspect ratio of the
underlying aura_surface. This aspect ratio includes the titlebar.
Unfortunately this logic (using a hardcoded 50 pixels) doesn't scale
correctly when resizing the window, leaving a blank space to the side
of the window.

This change sets aspect ratio using the actual viewport size, with
a hardcoded 32 pixels for the titlebar.

BUG=b:330639760
TEST=run containerized game, enter windowed mode via fullscreen key,
test resizing

Change-Id: If8c668f3ebaff5f2f33e7c9460f2990ada5bd125
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5545958
Commit-Queue: Arjun Srinivasan <asriniva@google.com>
Reviewed-by: Chloe Pelling <cpelling@google.com>
Tested-by: Arjun Srinivasan <asriniva@google.com>